third_party/blink/web_tests/accessibility/table-cells-with-colspan-expected.txt

data	col heada	data
data	col head
col head	col head

Check whether cells with colspan are reported properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS table1.columnCount is 3
PASS cell_0_0.role is 'AXRole: AXCell'
PASS cell_0_0.columnIndexRange() is '{0, 1}'
PASS cell_0_1.role is 'AXRole: AXColumnHeader'
PASS cell_0_1.columnIndexRange() is '{1, 1}'
PASS cell_0_2.role is 'AXRole: AXCell'
PASS cell_0_2.columnIndexRange() is '{2, 1}'
PASS table2.columnCount is 3
PASS table2.columnHeadersCount is 3
PASS cell_0_1.columnIndexRange() is '{1, 2}'
PASS cell_0_2.columnIndexRange() is '{1, 2}'
PASS table2.columnHeaderAtIndex(0).isEqual(cell_0_1) is true
PASS table2.columnHeaderAtIndex(0).isEqual(cell_0_2) is true
PASS table2.columnHeaderAtIndex(1).isEqual(table2.cellForColumnAndRow(1, 1)) is true
PASS table2.columnHeaderAtIndex(2).isEqual(table2.cellForColumnAndRow(2, 1)) is true
PASS successfullyParsed is true

TEST COMPLETE
